fix: 쿼리 키 수정 및 타입 임포트로 변경
